Aggregate values row-wise

Applies one or more provided summary functions row-wise to selected columns of a DTSg object.

# S3 method for DTSg
rowaggregate(
  x,
  resultCols,
  fun,
  ...,
  cols = self$cols(class = "numeric"),
  clone = getOption("DTSgClone")
)

Arguments

x
A DTSg object (S3 method only).

resultCols
A character vector either of length one (names of fun are appended in the case one or more functions are provided) or the same length as fun specifying the column names for the return values of fun.

fun
A summary function, (named) list of summary functions or (named) character vector specifying summary functions applied row-wise to all the values of the specified cols. The return value(s) must be of length one. See corresponding section for further information.

...
Further arguments passed on to fun.

cols
A character vector specifying the columns to apply fun to.

clone
A logical specifying if the object shall be modified in place or if a deep clone (copy) shall be made beforehand.

Value

Returns a DTSg object.

Summary functions

Some examples for fun are as follows:

mean

list(min = min, max = max)

c(sd = "sd", var = "var")

See also

Examples

# new DTSg object
DT <- data.table::data.table(
  date = flow$date,
  flow1 = flow$flow - rnorm(nrow(flow)),
  flow2 = flow$flow,
  flow3 = flow$flow + rnorm(nrow(flow))
)
x <- DTSg$new(values = DT)

# mean and standard deviation of multiple measurements per timestamp
## R6 method
x$rowaggregate(
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
)$print()

#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean   flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>     <num>
#>    1: 2007-01-01 10.940044  9.540  8.147391  9.542478 1.3963279
#>    2: 2007-01-02  9.029683  9.285  9.909948  9.408210 0.4528819
#>    3: 2007-01-03 11.377264  8.940  9.620720  9.979328 1.2575823
#>    4: 2007-01-04  8.750571  8.745 10.996140  9.497237 1.2980910
#>    5: 2007-01-05  7.868447  8.490  6.647879  7.668775 0.9371521
#>   ---                                                          
#> 2188: 2012-12-27 25.259652 26.685 25.721425 25.888692 0.7272469
#> 2189: 2012-12-28 27.514516 28.050 28.528658 28.031058 0.5073363
#> 2190: 2012-12-29 24.059124 23.580 24.360565 23.999896 0.3936386
#> 2191: 2012-12-30 18.131244 18.840 18.654575 18.541940 0.3675576
#> 2192: 2012-12-31 17.260941 17.250 16.878584 17.129842 0.2176641
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192

## S3 method
print(rowaggregate(
  x = x,
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
))

#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean   flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>     <num>
#>    1: 2007-01-01 10.940044  9.540  8.147391  9.542478 1.3963279
#>    2: 2007-01-02  9.029683  9.285  9.909948  9.408210 0.4528819
#>    3: 2007-01-03 11.377264  8.940  9.620720  9.979328 1.2575823
#>    4: 2007-01-04  8.750571  8.745 10.996140  9.497237 1.2980910
#>    5: 2007-01-05  7.868447  8.490  6.647879  7.668775 0.9371521
#>   ---                                                          
#> 2188: 2012-12-27 25.259652 26.685 25.721425 25.888692 0.7272469
#> 2189: 2012-12-28 27.514516 28.050 28.528658 28.031058 0.5073363
#> 2190: 2012-12-29 24.059124 23.580 24.360565 23.999896 0.3936386
#> 2191: 2012-12-30 18.131244 18.840 18.654575 18.541940 0.3675576
#> 2192: 2012-12-31 17.260941 17.250 16.878584 17.129842 0.2176641
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192