Get column names

Returns all column names of a DTSg object, those of certain classes, modes, typeofs and/or those matching a certain pattern only.

# S3 method for DTSg
cols(x, class = NULL, pattern = NULL, mode = NULL, typeof = NULL, ...)

Arguments

x

A DTSg object (S3 method only).

class

An optional character vector matched to the most specific class (first element) of each column's class vector. The “special class” ".numerary" matches the integer and numeric classes.

pattern

An optional character string passed on to the pattern argument of grep.

mode

An optional character vector matched to each column's mode vector.

typeof

An optional character vector matched to each column's typeof vector.

...

Further arguments passed on to grep. The value argument is rejected.

Value

Returns a character vector.

Examples

# new DTSg object
x <- DTSg$new(values = flow)

# get names of numeric columns
## R6 method
x$cols(class = "numeric")
#> [1] "flow"

## S3 method
cols(x = x, class = "numeric")
#> [1] "flow"