Aggregate values row-wise

Applies one or more provided summary functions row-wise to selected columns of a DTSg object.

# S3 method for DTSg
rowaggregate(
  x,
  resultCols,
  fun,
  ...,
  cols = self$cols(class = "numeric"),
  clone = getOption("DTSgClone")
)

Arguments

x

A DTSg object (S3 method only).

resultCols

A character vector either of length one (names of fun are appended in the case one or more functions are provided) or the same length as fun specifying the column names for the return values of fun.

fun

A summary function, (named) list of summary functions or (named) character vector specifying summary functions applied row-wise to all the values of the specified cols. The return value(s) must be of length one. See corresponding section for further information.

...

Further arguments passed on to fun.

cols

A character vector specifying the columns to apply fun to. Another possibility is a character string containing either comma separated column names, for example, "x,y,z", or the start and end column separated by a colon, for example, "x:z".

clone

A logical specifying if the object shall be modified in place or if a deep clone (copy) shall be made beforehand.

Value

Returns a DTSg object.

Summary functions

Some examples for fun are as follows:

mean

list(min = min, max = max)

c(sd = "sd", var = "var")

See also

Examples

# new DTSg object
DT <- data.table::data.table(
  date = flow$date,
  flow1 = flow$flow - abs(rnorm(nrow(flow))),
  flow2 = flow$flow,
  flow3 = flow$flow + abs(rnorm(nrow(flow)))
)
x <- DTSg$new(values = DT)

# mean and standard deviation of multiple measurements per timestamp
## R6 method
x$rowaggregate(
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
)$print()
#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean   flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>     <num>
#>    1: 2007-01-01  8.781386  9.540 10.531238  9.617541 0.8774992
#>    2: 2007-01-02  7.719722  9.285  9.713932  8.906218 1.0496787
#>    3: 2007-01-03  8.379530  8.940  9.024227  8.781252 0.3504417
#>    4: 2007-01-04  8.687597  8.745  9.544673  8.992423 0.4791230
#>    5: 2007-01-05  6.339430  8.490  9.840202  8.223210 1.7655689
#>   ---                                                          
#> 2188: 2012-12-27 26.095617 26.685 27.245812 26.675476 0.5751567
#> 2189: 2012-12-28 27.238614 28.050 29.351397 28.213337 1.0658201
#> 2190: 2012-12-29 22.964246 23.580 25.005115 23.849787 1.0468407
#> 2191: 2012-12-30 17.914260 18.840 19.265782 18.673347 0.6910013
#> 2192: 2012-12-31 17.249999 17.250 18.539944 17.679981 0.7447496
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192

## 'raggregate()' is a "hidden" R6 alias for 'rowaggregate()'
x$raggregate(
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
)$print()
#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean   flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>     <num>
#>    1: 2007-01-01  8.781386  9.540 10.531238  9.617541 0.8774992
#>    2: 2007-01-02  7.719722  9.285  9.713932  8.906218 1.0496787
#>    3: 2007-01-03  8.379530  8.940  9.024227  8.781252 0.3504417
#>    4: 2007-01-04  8.687597  8.745  9.544673  8.992423 0.4791230
#>    5: 2007-01-05  6.339430  8.490  9.840202  8.223210 1.7655689
#>   ---                                                          
#> 2188: 2012-12-27 26.095617 26.685 27.245812 26.675476 0.5751567
#> 2189: 2012-12-28 27.238614 28.050 29.351397 28.213337 1.0658201
#> 2190: 2012-12-29 22.964246 23.580 25.005115 23.849787 1.0468407
#> 2191: 2012-12-30 17.914260 18.840 19.265782 18.673347 0.6910013
#> 2192: 2012-12-31 17.249999 17.250 18.539944 17.679981 0.7447496
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192

## S3 method
print(rowaggregate(
  x = x,
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
))
#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean   flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>     <num>
#>    1: 2007-01-01  8.781386  9.540 10.531238  9.617541 0.8774992
#>    2: 2007-01-02  7.719722  9.285  9.713932  8.906218 1.0496787
#>    3: 2007-01-03  8.379530  8.940  9.024227  8.781252 0.3504417
#>    4: 2007-01-04  8.687597  8.745  9.544673  8.992423 0.4791230
#>    5: 2007-01-05  6.339430  8.490  9.840202  8.223210 1.7655689
#>   ---                                                          
#> 2188: 2012-12-27 26.095617 26.685 27.245812 26.675476 0.5751567
#> 2189: 2012-12-28 27.238614 28.050 29.351397 28.213337 1.0658201
#> 2190: 2012-12-29 22.964246 23.580 25.005115 23.849787 1.0468407
#> 2191: 2012-12-30 17.914260 18.840 19.265782 18.673347 0.6910013
#> 2192: 2012-12-31 17.249999 17.250 18.539944 17.679981 0.7447496
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192