Aggregate values row-wise

Applies one or more provided summary functions row-wise to selected columns of a DTSg object.

# S3 method for DTSg
rowaggregate(
  x,
  resultCols,
  fun,
  ...,
  cols = self$cols(class = "numeric"),
  clone = getOption("DTSgClone")
)

Arguments

x

A DTSg object (S3 method only).

resultCols

A character vector either of length one (names of fun are appended in the case one or more functions are provided) or the same length as fun specifying the column names for the return values of fun.

fun

A summary function, (named) list of summary functions or (named) character vector specifying summary functions applied row-wise to all the values of the specified cols. The return value(s) must be of length one. See corresponding section for further information.

...

Further arguments passed on to fun.

cols

A character vector specifying the columns to apply fun to. Another possibility is a character string containing either comma separated column names, for example, "x,y,z", or the start and end column separated by a colon, for example, "x:z".

clone

A logical specifying if the object shall be modified in place or if a deep clone (copy) shall be made beforehand.

Value

Returns a DTSg object.

Summary functions

Some examples for fun are as follows:

mean

list(min = min, max = max)

c(sd = "sd", var = "var")

See also

Examples

# new DTSg object
DT <- data.table::data.table(
  date = flow$date,
  flow1 = flow$flow - abs(rnorm(nrow(flow))),
  flow2 = flow$flow,
  flow3 = flow$flow + abs(rnorm(nrow(flow)))
)
x <- DTSg$new(values = DT)

# mean and standard deviation of multiple measurements per timestamp
## R6 method
x$rowaggregate(
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
)$print()
#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean    flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>      <num>
#>    1: 2007-01-01  9.053088  9.540 11.423210 10.005433 1.25173510
#>    2: 2007-01-02  8.011010  9.285 11.381362  9.559124 1.70181541
#>    3: 2007-01-03  8.812572  8.940  9.152214  8.968262 0.17157592
#>    4: 2007-01-04  8.722314  8.745 10.615989  9.361101 1.08682385
#>    5: 2007-01-05  7.080257  8.490  9.517196  8.362484 1.22346349
#>   ---                                                           
#> 2188: 2012-12-27 26.651916 26.685 26.735222 26.690712 0.04194582
#> 2189: 2012-12-28 27.961851 28.050 28.511148 28.174333 0.29500119
#> 2190: 2012-12-29 23.359272 23.580 24.343855 23.761042 0.51665583
#> 2191: 2012-12-30 18.340645 18.840 19.059667 18.746771 0.36846554
#> 2192: 2012-12-31 15.398722 17.250 17.666021 16.771581 1.20699001
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192

## S3 method
print(rowaggregate(
  x = x,
  resultCols = "flow",
  fun = list(mean = mean, sd = sd)
))
#> Values:
#>        .dateTime     flow1  flow2     flow3 flow.mean    flow.sd
#>           <POSc>     <num>  <num>     <num>     <num>      <num>
#>    1: 2007-01-01  9.053088  9.540 11.423210 10.005433 1.25173510
#>    2: 2007-01-02  8.011010  9.285 11.381362  9.559124 1.70181541
#>    3: 2007-01-03  8.812572  8.940  9.152214  8.968262 0.17157592
#>    4: 2007-01-04  8.722314  8.745 10.615989  9.361101 1.08682385
#>    5: 2007-01-05  7.080257  8.490  9.517196  8.362484 1.22346349
#>   ---                                                           
#> 2188: 2012-12-27 26.651916 26.685 26.735222 26.690712 0.04194582
#> 2189: 2012-12-28 27.961851 28.050 28.511148 28.174333 0.29500119
#> 2190: 2012-12-29 23.359272 23.580 24.343855 23.761042 0.51665583
#> 2191: 2012-12-30 18.340645 18.840 19.059667 18.746771 0.36846554
#> 2192: 2012-12-31 15.398722 17.250 17.666021 16.771581 1.20699001
#> 
#> Aggregated:     FALSE
#> Regular:        TRUE
#> Periodicity:    Time difference of 1 days
#> Missing values: explicit
#> Time zone:      UTC
#> Timestamps:     2192